fix another memory leak
This memory leak showed up on valgrind when `rc_get_next_id_str_callback` is called with `rc_got_users_of_room`. I am aware that the `room_id` was supposed to be freed in `rc_got_users_of_room` but apparently that didn't always happen. Therefore I introduced this `user_data_destroy_func` so the `user_data` is always freed when the item is removed from the hash table.

For consistency the `user_data_destroy_func` is also set for the `rc_got_history_of_room` case although there was no leak showing up when I ran this for a short while in valgrind.