Only start the actual login process when api/me could be
Wait for the call to `<server>/api/me` was successful until we start
the websocket and register additional signals.

Doing so is almost the same as actually checking if the server is a
rocket server as the call to `<server>/api/me` basically means that.

Add todo for to improve debug output.

Signed-off-by: Björn Bidar <bjorn.bidar@thaodan.de>